requirement.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==2.2.9
django-filter==2.3.0
django-jet==1.0.8
psycopg2-binary==2.8.5
python-http-client==3.3.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
django-reversion==3.0.4